Normalize CRS handling (#2116)
* Normalize CRS handling

* Normalize storage_crs

* Update crs.py

* Cleanup

* Use CrsTransformSpec for all CRS conversion

* Create `get_transform_from_spec`

* Add additional tests

* Reduce diff

* Revert transform_bbox changes

* Update documentation

* Update CRS docs

* Update plugins.rst

* Fix comment spacing in configuration docs

* Incorporate PR review feedback

* Small change to verbiage
