+++ pygeoapi/provider/sqlite.py
@@ -35,10 +35,11 @@ import sqlite3
 import logging
 import os
 import json
+
+from pygeoapi.crs import crs_transform
 from pygeoapi.plugin import InvalidPluginError
 from pygeoapi.provider.base import (BaseProvider, ProviderConnectionError,
                                     ProviderItemNotFoundError)
-from pygeoapi.util import crs_transform
 
 LOGGER = logging.getLogger(__name__)
